Initial model magic code.
This code is the initial relationship building code for sets, it creates the set model and
builds the relationships about the involved models.

Parent 1:N Set N:N Child

The Set/Child relationship is build a regular many-to-many relationship with Rails, the Parent/Set relationship is built using belongs_to, that is, the parent has to have an FK to the newly generated Set class.

This code assumes can't be run as-is unless you have the migrations and parent/child models ready, generators do not exist yet so this has to be done by hand. The SQL code to handle the Sets behind the curtains is not there yet as well, this is just the magic relationship code.