Added a generator for the sets migration.
Sets need two parameters, the parent table and the child (related) table. What this
migration will create is the set table, a regular many-to-many relationship table between
the set table and the child table, and will add a column to relate the parent column to
the set table hence creating the full many-to-many relationship. Indexes are added for
all relationship columns (foreign keys).