Setter method for set items is a bit smarter.
The behavior is a bit more smarter to work nicely with Rails ORM operations. If an
element is set that is not an array but can be converted (like ActiveRecord::Relation
which is what .where() returns) it will be converted. If not an Array and cannot be
converted, then it will be wrapped inside an array and hope for the best.