Removed preloading of sets.
There's no point in preloading since the items may not get used and it increases the
risk of racing conditions. Now we have a private loader method that gets the data from
the underlying relationship that is used by the getter, and now the saving code uses the
getter as well instead of going for the instance variable, is safer.