Fixed a bug when saving set items, we could save duplicates.
It is possible to add duplicates entries to the set items array because we return its
reference so people can use things like << and it behaves in the less surprising manner.
The setter now runs uniq along with flatten to make the values behave as we expect, also,
when saving the save code does an extra check on this as well and sets the current values
to the ones actually saved.

Added a test for this behavior, it didnt pass before, with the code in this commit it does.