Children can now see their parents.
Added a way for a child model to see to which parents it belongs to, in a similar fashion
to what has_and_belongs_to_many does. This is just for reading though, even if the result
is modified this won't go into the database at all. Then the parents are requested, the
query gets run and the results get cached for subsequent uses.

A couple of tests were added to exercise this functionality.